% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_library.R
\name{build_library_cache}
\alias{build_library_cache}
\title{Build design library template and cache diagnosis grid}
\usage{
build_library_cache(topic, file = file.path("~/cache", paste0(topic,
  ".Rdata")), cache_loc = file.path("~/cache", topic))
}
\arguments{
\item{topic}{the vignette topic}

\item{file}{to save the Rdata to}

\item{cache_loc}{to save the cache data to}
}
\description{
Build design library template and cache diagnosis grid
}
