% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{dd_install_github}
\alias{dd_install_github}
\title{Install all DD packages}
\usage{
dd_install_github(DeclareDesign = "master", shiny = "master",
  randomizr = "master", fabricatr = "master", estimatr = "master", ...)
}
\arguments{
\item{DeclareDesign}{branch or tag}

\item{shiny}{branch or tag}

\item{randomizr}{branch or tag}

\item{fabricatr}{branch or tag}

\item{estimatr}{branch or tag}
}
\description{
Install all DD packages
}
