% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{enable_parallel_make}
\alias{enable_parallel_make}
\title{Enable parallel make}
\usage{
enable_parallel_make(Makevar = file.path("~", ".R", "Makevars"))
}
\arguments{
\item{Makevar}{the Makevar file to update}
}
\description{
This enables building with 8 cores by setting \code{MAKE=make -j8}
}
