% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{plot_deps}
\alias{plot_deps}
\title{Make dependency plot}
\usage{
plot_deps(fields = c("Depends", "Imports", "LinkingTo", "Suggests"))
}
\arguments{
\item{fields}{which fields of the description to use}
}
\description{
Make dependency plot
}
