% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paren_hack.R
\name{print.paren_hack}
\alias{print.paren_hack}
\title{Paren hack}
\usage{
\method{print}{paren_hack}(x, ...)
}
\description{
By adding the paren_hack class to a function, you can invoke it via print -
this means you don't need ()s to do so.
}
